requirements.txt
anyio==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
psycopg2-binary
pyTelegramBotAPI==4.2.0
requests==2.26.0
sniffio==1.2.0
starlette==0.17.0
urllib3==1.26.7
